requirements.txt
django==3.0.5
gunicorn==20.0.4
django-ckeditor==5.9.0
django-constance==2.6.0
django-picklefield==2.1.1